Added option to be regular menu and (hopefully) a protection against not being able to write a new file
Previously if the storage media was full it would end up erasing the menu without replacing it, so the user would have no menu.
Now it will first write the new menu file to a temporary file, then try to check if the write was successful before deleting the old menu and renaming the new menu. I didn't test it, so I can only hope it works.
This doesn't work in python 2.7, so to maintain compatibility I check the python version. The old bevviour is maintained for python 2.7.