Added default locations to search for files.
For config: first in the file locaion provided in the arguments. If no file was specified, then in the user's config. If it doesn't exist there, then system config.
For cache: first in the provided location. If cache wasn't specified in the config, then in the user's config.
For header and footer: in the specified location according to config.

For all of the above: support was added to checking as a relative path to the config file's directory if finding as absolute path fails.